requirements.txt
amqp==5.0.6
asgiref==3.3.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
celery==5.0.5
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.1.6
django-debug-toolbar==3.2.2
django-environ==0.4.5
django-extensions==3.1.3
django-redis==5.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.0
environ==1.0
idna==2.10
kombu==5.1.0
Pillow==8.2.0
pkg_resources==0.0.0
prompt-toolkit==3.0.19
psycopg2==2.9.1
PyJWT==2.4.0
python-decouple==3.4
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
